+++ src/components/molecules/ArticleList/index.vue
@@ -81,37 +81,6 @@
         削除
       </app-button>
     </app-modal>
-    <div class="pagination">
-      <app-button
-        class="pagination__button"
-        :disabled="currentPage === 1"
-        @click="goToPage(1)"
-      >
-        1
-      </app-button>
-      <span class="pagination__span">...</span>
-      <app-button
-        v-for="page in generatePageNumbers()"
-        :key="page"
-        class="pagination__button"
-        :disabled="currentPage === page"
-        @click="goToPage(page)"
-      >
-        {{ page }}
-      </app-button>
-      <span
-        v-if="totalPages > 1"
-        class="pagination__span"
-      >...</span>
-      <app-button
-        v-if="totalPages > 1"
-        class="pagination__button"
-        :disabled="currentPage === totalPages"
-        @click="goToPage(totalPages)"
-      >
-        {{ totalPages }}
-      </app-button>
-    </div>
   </div>
 </template>
 
@@ -157,18 +126,6 @@ export default {
       type: Object,
       default: () => ({}),
     },
-    currentPage: {
-      type: Number,
-      default: null,
-    },
-    totalPages: {
-      type: Number,
-      default: null,
-    },
-    visiblePages: {
-      type: Number,
-      default: 5,
-    },
   },
   computed: {
     articleTitle() {
@@ -182,50 +139,6 @@ export default {
     openModal(articleId) {
       if (!this.access.delete) return;
       this.$emit('open-modal', articleId);
-    },
-    goToPage(pageNumber) {
-      if (pageNumber === this.currentPage) {
-        return;
-      }
-      this.$emit('go-to-page', pageNumber);
-    },
-    isFirstPageVisible() {
-      return this.currentPage > Math.floor(this.visiblePages / 2) + 1;
-    },
-    isLastPageVisible() {
-      return this.totalPages - this.currentPage >= Math.floor(this.visiblePages / 2);
-    },
-    generatePageNumbers() {
-      const pageNumbers = [];
-      let startPage = 2;
-      let endPage = this.totalPages - 1;
-
-      if (this.totalPages - 1 > this.visiblePages) {
-        const maxVisiblePages = this.visiblePages;
-        const offset = Math.floor(maxVisiblePages / 2);
-
-        if (this.isFirstPageVisible()) {
-          startPage = this.currentPage - offset;
-          endPage = this.currentPage + offset;
-
-          if (endPage > this.totalPages - 1) {
-            endPage = this.totalPages - 1;
-            startPage = endPage - maxVisiblePages + 1;
-          }
-        } else if (this.isLastPageVisible()) {
-          endPage = this.currentPage + offset;
-          startPage = endPage - maxVisiblePages + 1;
-
-          if (startPage < 2) {
-            startPage = 2;
-            endPage = startPage + maxVisiblePages - 1;
-          }
-        }
-      }
-      for (let i = startPage; i <= endPage; i += 1) {
-        pageNumbers.push(i);
-      }
-      return pageNumbers;
     },
   },
 };
@@ -255,22 +168,6 @@ export default {
     }
     &__notice--create {
       margin-bottom: 16px;
-    }
-  }
-
-  .pagination {
-    margin-top: 30px;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    &__button {
-      margin: 0 10px 0;
-    }
-    &__span {
-      margin: 0 15px;
-      font-size: 22px;
-      opacity: 0.5;
     }
   }
 </style>
