+++ src/components/molecules/index.js
@@ -16,6 +16,7 @@ import ArticlePost from './ArticlePost/index.vue';
 import ArticleDetail from './ArticleDetail/index.vue';
 import DeleteModal from './Modal/Delete.vue';
 import Notice from './Notice/index.vue';
+import Pagination from './Pagination/index.vue';
 
 export {
   SigninForm,
@@ -36,4 +37,5 @@ export {
   ArticleDetail,
   DeleteModal,
   Notice,
+  Pagination,
 };
